[Docs]: Change component name
